security/manager/ssl/tests/mochitest/mixedcontent/test_innerHtmlDelayedUnsecurePicture.html

<!DOCTYPE HTML>
<html>
<head>
  <title>innerHTML changes to unsecure test</title>
  <script type="text/javascript" src="/MochiKit/Base.js"></script>
  <script type="text/javascript" src="/MochiKit/DOM.js"></script>
  <script type="text/javascript" src="/MochiKit/Style.js"></script>
  <script type="text/javascript" src="/MochiKit/Signal.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="mixedContentTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />

  <script class="testbody" type="text/javascript">

  function runTest()
  {
    isSecurityState("secure");
    
    window.setTimeout(function() 
    {
      document.getElementById("buddy").innerHTML =
        "<img id='image1' src='http://example.com/tests/security/manager/ssl/tests/mochitest/mixedcontent/moonsurface.jpg' />";
    }, 1);

    waitForSecurityState("broken", function() 
    {
      isSecurityState("broken", "innerHTML loading insecure changed to broken");
      finish();
    });
  }

  function afterNavigationTest()
  {
    is(document.getElementById("buddy").innerHTML, "", "innerHTML back to previous");
    isSecurityState("secure");
    finish();    
  }

  </script>
</head>

<body id="buddy"></body></html>
